+++ objects/src/accounts/mod.rs
@@ -18,7 +18,10 @@ pub mod code;
 pub use code::{procedure::AccountProcedureInfo, AccountCode};
 
 pub mod delta;
-pub use delta::{AccountDelta, AccountStorageDelta, AccountVaultDelta, StorageMapDelta};
+pub use delta::{
+    AccountDelta, AccountStorageDelta, AccountVaultDelta, FungibleAssetDelta,
+    NonFungibleAssetDelta, NonFungibleDeltaAction, StorageMapDelta,
+};
 
 mod seed;
 pub use seed::{get_account_seed, get_account_seed_single};
@@ -189,15 +192,11 @@ impl Account {
     /// - The nonce specified in the provided delta smaller than or equal to the current account
     ///   nonce.
     pub fn apply_delta(&mut self, delta: &AccountDelta) -> Result<(), AccountError> {
-        // update vault; we don't check vault delta validity here because AccountDelta can contain
+        // update vault; we don't check vault delta validity here because `AccountDelta` can contain
         // only valid vault deltas
-        for &asset in delta.vault().added_assets.iter() {
-            self.vault.add_asset(asset).map_err(AccountError::AssetVaultUpdateError)?;
-        }
-
-        for &asset in delta.vault().removed_assets.iter() {
-            self.vault.remove_asset(asset).map_err(AccountError::AssetVaultUpdateError)?;
-        }
+        self.vault
+            .apply_delta(delta.vault())
+            .map_err(AccountError::AssetVaultUpdateError)?;
 
         // update storage
         self.storage.apply_delta(delta.storage())?;
@@ -323,10 +322,10 @@ mod tests {
 
     use super::{AccountDelta, AccountStorageDelta, AccountVaultDelta};
     use crate::{
-        accounts::{
-            delta::AccountStorageDeltaBuilder, Account, SlotItem, StorageMap, StorageMapDelta,
+        accounts::{Account, SlotItem, StorageMap, StorageMapDelta},
+        testing::storage::{
+            build_account, build_account_delta, build_assets, AccountStorageDeltaBuilder,
         },
-        testing::storage::{build_account, build_account_delta, build_assets},
     };
 
     #[test]
@@ -346,7 +345,7 @@ mod tests {
     fn test_serde_account_delta() {
         let final_nonce = Felt::new(2);
         let (asset_0, asset_1) = build_assets();
-        let storage_delta = AccountStorageDeltaBuilder::new()
+        let storage_delta = AccountStorageDeltaBuilder::default()
             .add_cleared_items([0])
             .add_updated_items([(1_u8, [Felt::new(1), Felt::new(2), Felt::new(3), Felt::new(4)])])
             .build()
@@ -397,13 +396,13 @@ mod tests {
             [Felt::new(9_u64), Felt::new(10_u64), Felt::new(11_u64), Felt::new(12_u64)],
         );
         let updated_map =
-            StorageMapDelta::from(vec![], vec![(new_map_entry.0.into(), new_map_entry.1)]);
+            StorageMapDelta::from_iters([], [(new_map_entry.0.into(), new_map_entry.1)]);
         storage_map.insert(new_map_entry.0, new_map_entry.1);
         maps.insert(2u8, storage_map.clone());
 
         // build account delta
         let final_nonce = Felt::new(2);
-        let storage_delta = AccountStorageDeltaBuilder::new()
+        let storage_delta = AccountStorageDeltaBuilder::default()
             .add_cleared_items([0])
             .add_updated_items([(1_u8, [Felt::new(1), Felt::new(2), Felt::new(3), Felt::new(4)])])
             .add_updated_maps([(2_u8, updated_map)])
@@ -444,7 +443,7 @@ mod tests {
         );
 
         // build account delta
-        let storage_delta = AccountStorageDeltaBuilder::new()
+        let storage_delta = AccountStorageDeltaBuilder::default()
             .add_cleared_items([0])
             .add_updated_items([(1_u8, [Felt::new(1), Felt::new(2), Felt::new(3), Felt::new(4)])])
             .build()
@@ -470,7 +469,7 @@ mod tests {
 
         // build account delta
         let final_nonce = Felt::new(1);
-        let storage_delta = AccountStorageDeltaBuilder::new()
+        let storage_delta = AccountStorageDeltaBuilder::default()
             .add_cleared_items([0])
             .add_updated_items([(1_u8, [Felt::new(1), Felt::new(2), Felt::new(3), Felt::new(4)])])
             .build()